Move RingbufferLogger construction
